Sport class to new file
