+++ Data_Download/Sports_Data/formula-1/Status/Status.json
@@ -19,7 +19,7 @@
             "active": true
         },
         "requests": {
-            "current": 1,
+            "current": 2,
             "limit_day": 100
         }
     }
