+++ Data_Download/Sports_Data/nba/Status/Status.json
@@ -19,7 +19,7 @@
             "active": true
         },
         "requests": {
-            "current": 2,
+            "current": 4,
             "limit_day": 100
         }
     }
